Reorganized binary_galprop_model into the component_model_templates sub-package
