+++ halotools/empirical_models/component_model_templates/tests/test_binary_galprop_models.py
@@ -9,7 +9,7 @@ from ... import model_defaults
 
 def test_BinaryGalpropInterpolModel():
     """ Function testing the initialization of 
-    `~halotools.empirical_models.sfr_components.BinaryGalpropInterpolModel`. 
+    `~halotools.empirical_models.BinaryGalpropInterpolModel`. 
     """
     abcissa, ordinates = [12, 15], [1/3., 0.9]
     m = BinaryModel(galprop_name='late_type',
