+++ halotools/empirical_models/sfr_models/__init__.py
@@ -1,1 +1,1 @@
-from .sfr_components import *
+from .halo_mass_quenching import *
